Move VM from package vm_test to vm, because it's not tests.
Now the vm_test package is free for tests *of* the VM.
