+++ actors/test/cron_catches_expiries_scenario_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/filecoin-project/specs-actors/v5/actors/runtime/proof"
 	"github.com/filecoin-project/specs-actors/v5/support/ipld"
 	tutil "github.com/filecoin-project/specs-actors/v5/support/testing"
-	vm "github.com/filecoin-project/specs-actors/v5/support/vm"
+	"github.com/filecoin-project/specs-actors/v5/support/vm"
 )
 
 var fakeChainRandomness = []byte("not really random")
