+++ actors/test/power_scenario_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/filecoin-project/specs-actors/v5/actors/builtin/power"
 	"github.com/filecoin-project/specs-actors/v5/support/ipld"
 	tutil "github.com/filecoin-project/specs-actors/v5/support/testing"
-	vm "github.com/filecoin-project/specs-actors/v5/support/vm"
+	"github.com/filecoin-project/specs-actors/v5/support/vm"
 )
 
 func TestCreateMiner(t *testing.T) {
